requirements.txt
django==3.1.7
django-cors-headers==3.2.1
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11
djangorestframework-simplejwt==4.6.0
dropbox==11.0.0
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
geoip2==4.1.0